+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugins/_internal/ui/content/DeveloperToolContentPanel.kt
@@ -10,6 +10,7 @@ import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.ui.Messages.InputDialog
 import com.intellij.ui.ScrollPaneFactory.createScrollPane
 import com.intellij.ui.dsl.builder.Align
+import com.intellij.ui.dsl.builder.Row
 import com.intellij.ui.dsl.builder.panel
 import com.intellij.ui.tabs.JBTabs
 import com.intellij.ui.tabs.JBTabsFactory
@@ -26,8 +27,8 @@ import dev.turingcomplete.intellijdevelopertoolsplugins._internal.ui.menu.Develo
 import javax.swing.Icon
 import javax.swing.JComponent
 
-internal class DeveloperToolContentPanel(
-  private val developerToolNode: DeveloperToolNode
+internal open class DeveloperToolContentPanel(
+  protected val developerToolNode: DeveloperToolNode
 ) : BorderLayoutPanel() {
   // -- Properties -------------------------------------------------------------------------------------------------- //
 
@@ -51,15 +52,19 @@ internal class DeveloperToolContentPanel(
     selectedDeveloperToolInstance.get().instance.deactivated()
   }
 
+  @Suppress("DialogTitleCapitalization")
+  protected open fun Row.buildTitle() {
+    label(developerToolNode.developerToolPresentation.contentTitle)
+      .applyToComponent { font = JBFont.label().asBold() }
+      .align(Align.FILL)
+      .resizableColumn()
+  }
+
   // -- Private Methods --------------------------------------------------------------------------------------------- //
 
-  @Suppress("DialogTitleCapitalization")
   private fun createTitleBar(): JComponent = panel {
     row {
-      label(developerToolNode.developerToolPresentation.contentTitle)
-        .applyToComponent { font = JBFont.label().asBold() }
-        .align(Align.FILL)
-        .resizableColumn()
+      buildTitle()
 
       link("Reset") {
         selectedDeveloperToolInstance.get().apply {